refactor: move app whitelabel settings to config.js
